requirements.bluefin.txt
# AI Trading Bot - Bluefin Requirements (Fixed Socket.IO)
# Generated for UV installation with Bluefin SDK

# Core dependencies
python-dotenv>=1.0.1
pydantic>=2.10.5
pydantic-settings>=2.7.0
click>=8.1.7

# Data and Analysis - Install FIRST to avoid conflicts
pandas>=2.3.0
numpy>=1.26.4
pandas-ta>=0.3.14b
scipy>=1.15.3

# LangChain and OpenAI
langchain>=0.3.10
langchain-core>=0.3.22
langchain-openai>=0.2.14
openai>=1.58.1

# Trading and Finance  
coinbase-advanced-py>=1.6.4

# Async and Networking
aiohttp==3.8.6
aiofiles>=24.1.0
websockets>=12.0
websocket-client==1.6.4
httpx>=0.28.1
backoff>=2.2.1

# Socket.IO - compatible versions for Bluefin SDK (exclude conflicting packages)
python-socketio>=5.8.0,<6.0.0
python-engineio>=4.6.0,<4.7.0

# Utilities
psutil>=6.1.0
docker>=7.1.0

# UI and Presentation
rich>=13.9.4

# Bluefin DEX SDK
bluefin-v2-client==3.2.4